add virtual joypad to componente
